refactor: 필요없는 코드 삭제, 패키지 정리
