--- asset/아아.txt
@@ -1,1 +1,0 @@
-rkskrkkk
