--- src/main/java/org/sopt/utils/IdGenerator.java
@@ -1,9 +1,0 @@
-package org.sopt.utils;
-
-public class IdGenerator {
-    private static int postId = 1;
-
-    public static int generatePostId() {
-        return postId++;
-    }
-}
